Moving file to components
