+++ src/plugin/components/StatusIcons.vue
